added user authorization specific things

+++ myshop/views.py
@@ -2,8 +2,9 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework import status
 from django.http import Http404
+from rest_framework.permissions import IsAuthenticated
 from .models import Product, Comment, CartItem, Category, ProductPicture, OrderItem, Order
-from .serializers import ProductSerializer, CommentSerializer, CartItemSerializer, CategorySerializer, CartSerializer, ProductCreateSerializer, PictureSerializer, CommentDetailSerializer, OrderItemSerializer, OrderSerializer
+from .serializers import ProductSerializer, CommentSerializer, CartItemSerializer, CategorySerializer, CartSerializer, ProductCreateSerializer, PictureSerializer, CommentDetailSerializer, OrderItemSerializer, OrderSerializer, CommentPatchSerializer, CartPatchSerializer
 from django.contrib.auth.models import User
 
 
@@ -23,7 +24,7 @@ class ProductsListView(APIView):
     def post(self, request):
         serializer = ProductCreateSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save()
+            serializer.save(supplier=request.user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
@@ -36,11 +37,15 @@ class ProductDetailView(APIView):
 
     def delete(self, request, pk):
         product = get_object(Product, pk)
+        if product.supplier != request.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
     def patch(self, request, pk):
         product = get_object(Product, pk)
+        if product.supplier != request.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
         serializer = ProductCreateSerializer(product, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
@@ -55,6 +60,8 @@ class ProductPicturesListView(APIView):
         return Response(serializer.data)
 
     def post(self, request, pk):
+        if Product.objects.get(id=pk).supplier != request.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
         picture = request.data['picture']
         ProductPicture.objects.create(product=Product.objects.get(id=pk), picture=picture)
         return Response(status=status.HTTP_201_CREATED)
@@ -62,11 +69,13 @@ class ProductPicturesListView(APIView):
 
 class ProductPictureDetailView(APIView):
     def get(self, request, pk, alt_pk):
-        picture = ProductPicture.objects.get(id=alt_pk)
+        picture = get_object(ProductPicture, alt_pk)
         serializer = PictureSerializer(picture)
         return Response(serializer.data)
 
     def delete(self, request, pk, alt_pk):
+        if Product.objects.get(id=pk).supplier != request.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
         picture = ProductPicture.objects.get(id=alt_pk)
         picture.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
@@ -113,10 +122,9 @@ class CommentsView(APIView):
     def post(self, request, pk):
         if request.data.get('rate') and request.data.get('comment_of_reply'):
             return Response(status=status.HTTP_400_BAD_REQUEST)
-        request.data['product'] = pk 
         serializer = CommentDetailSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save()
+            serializer.save(client=request.user, product=Product.objects.get(id=pk))
             print(serializer)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
@@ -128,20 +136,21 @@ class CommentDetailView(APIView):
         return Response(serializer.data)
 
     def delete(self, request, pk, alt_pk):
+        if Comment.objects.get(id=alt_pk).client != request.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
         comment = get_object(Comment, alt_pk)
         comment.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
     def patch(self, request, pk, alt_pk):
+        if Comment.objects.get(id=alt_pk).client != request.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
+
         comment = get_object(Comment, alt_pk)
-        if comment.comment_of_reply and request.data['rate']:
+        if comment.comment_of_reply and request.data.get('rate'):
             return Response(status=status.HTTP_400_BAD_REQUEST)
 
-        for key in list(request.data.keys()):
-            if key not in ['rate', 'content']:
-                request.data.pop(key)
-
-        serializer = CommentDetailSerializer(comment, data=request.data, partial=True)
+        serializer = CommentPatchSerializer(comment, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
@@ -149,15 +158,14 @@ class CommentDetailView(APIView):
 
 class CartView(APIView):
     def get(self, request):
-        cart_items = CartItem.objects.all()
+        cart_items = CartItem.objects.filter(client=request.user)
         serializer = CartSerializer({'products':cart_items})
         return Response(serializer.data)
 
     def post(self, request):
-        print(request.data)
         serializer = CartItemSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save()
+            serializer.save(client=request.user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)          
 
@@ -165,21 +173,25 @@ class CartView(APIView):
 class CartDetailView(APIView):
     def get(self, request, pk):
         cart_item = get_object(CartItem, pk)
+
+        if CartItem.objects.get(id=pk).client != request.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
+
         serializer = CartItemSerializer(cart_item)
         return Response(serializer.data)
 
     def delete(self, request, pk):
         cart_item = get_object(CartItem, pk)
+        if CartItem.objects.get(id=pk).client != request.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
         cart_item.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
-    def patch(self, request, pk):
-        for key in list(request.data.keys()):
-            if key not in ['quantity', 'status']:
-                request.data.pop(key)
-    
+    def patch(self, request, pk):    
         cart_item = get_object(CartItem, pk)
-        serializer = CartItemSerializer(cart_item, data=request.data, partial=True)
+        if CartItem.objects.get(id=pk).client != request.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
+        serializer = CartPatchSerializer(cart_item, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
